src/github_deps_tracker/pyproject.toml
[project]
name = "github-deps-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dlt[s3]>=1.24.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "dagster",
    "dagster-webserver",
    "dbt-clickhouse",
    "dagster-dbt",
]
